WARNING: this code is rather broken...
This is not much more than a proof of concept of how we could be using
LWP::Parallel::RobotUA for our requests. The basic idea is to send the
list of links as an array to $ua->register() instead of request()ing
them sequentially.

Done by
- modifying check_uri to pass an array of URIs to check_validity
- modifying check_validity to pass this array to  get_uris (a clone of get_uri)
- cloning get_uri into get_uris

The latter is only roughly done, and there probably is some loss of
information (recursion, etc) in the process.

This is a rather poor patch but the best I can wrap my head around right now,
feel free to revert if it's too ugly.